Chuck-A-Luck Simulations
Imagine letting 100 gamblers enter the casino with $2 each and play rounds of
Chuck-a-Luck. At the beginning of every round of the game, each player invests
exactly $1 (if he has the money). He continues to play round after round of the game until he either loses
everything, or his wealth reaches at least $4.
This code returns the average number of rounds the 100 gamblers played and the proportion of gamblers who
finished with at least $4 in their pocket as well as histogram of the number of rounds played for the players
who went bankrupt only, the histogram of the number of rounds played for the remaining players, and the
histogram of the number of rounds played for all players.